pyproject.toml
[tool.poetry]
name = "django-demo"
version = "0.5.0"
description = ""
authors = ["hsiehpinghan <thank.hsiehpinghan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "5.0.1"
djangorestframework = "3.14.0"
langchain-huggingface = "0.0.3"
langchain-openai = "0.1.10"
langchain = "0.2.7"
python-dotenv = "1.0.1"
faiss-cpu = "1.7.4"
langchain-community = "0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
